+++ src/components/Table/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Table'
 export * from './types'
-export * from './Table'
+export * from './helpers'
