+++ src/components/Table/views/CardView/index.ts
@@ -1,1 +1,1 @@
-export { default } from './CardView'
+export { CardView } from './CardView'
