+++ src/components/Table/views/TableView/TableView.tsx
@@ -1,38 +1,102 @@
-import { Table as AntdTable } from 'antd'
+import React, { ForwardedRef, ReactNode, Ref, forwardRef, memo, useMemo } from 'react'
 
-import { TableProps } from '../../Table'
+import classNames from 'classnames'
+import { TableVirtuoso } from 'react-virtuoso'
+
 import { getCardOrRowClassName } from '../../helpers'
+import { TableRowParams, TableViewProps } from '../../types'
 
-type TableViewProps<T> = Omit<TableProps<T, null>, 'sortViewParams' | 'loading'>
+import styles from './TableView.module.less'
 
-const TableView = <T extends object>({
+const TableViewInner = <T extends object>({
   data,
   className,
   columns,
-  onRowClick,
-  activeRowParams,
-  rowKeyField,
-  scrollX = 768,
+  rowParams,
+  loadMore,
 }: TableViewProps<T>) => {
-  const rowProps = (rowData: T) => ({
-    onClick: onRowClick ? () => onRowClick(rowData) : undefined,
-    style: onRowClick && { cursor: 'pointer' },
-  })
+  const rowProps = {
+    onClick: rowParams?.onRowClick ? (rowData: T) => rowParams?.onRowClick?.(rowData) : () => null,
+    style: rowParams?.onRowClick && { cursor: 'pointer' },
+  }
+
+  const tableComponents = useMemo(() => {
+    return createTableComponents(rowParams)
+  }, [rowParams])
 
   return (
-    <AntdTable
-      rowKey={(record) => record[rowKeyField] as string}
-      dataSource={[...data]}
-      columns={columns}
-      className={className}
-      rowClassName={(record) => getCardOrRowClassName(record, activeRowParams)}
-      rootClassName="rootTableClassName"
-      sortDirections={['descend', 'ascend']}
-      onRow={rowProps}
-      pagination={false}
-      scroll={{ x: scrollX }}
+    <TableVirtuoso
+      data={data}
+      overscan={200}
+      endReached={loadMore}
+      className={classNames(styles.tableWrapper, className)}
+      fixedHeaderContent={() => (
+        <tr>
+          {columns.map(({ key, title }, index) => {
+            return (
+              <th key={key} align={index ? 'right' : 'left'}>
+                {title}
+              </th>
+            )
+          })}
+        </tr>
+      )}
+      components={tableComponents}
+      itemContent={(index) => (
+        <>
+          {columns.map(({ key, render }) => {
+            return (
+              <td key={`${key}-${index}`} style={rowProps.style} align="right">
+                {render?.(data[index], index) as ReactNode}
+              </td>
+            )
+          })}
+        </>
+      )}
     />
   )
 }
 
-export default TableView
+export const TableView = memo(TableViewInner) as typeof TableViewInner
+
+interface TableRowProps<T>
+  extends React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
+  'data-index': number
+  item: T
+  rowParams?: TableRowParams<T>
+}
+
+function TableRowInner<T>(
+  props: TableRowProps<T>,
+  ref: ForwardedRef<HTMLTableRowElement>,
+): JSX.Element {
+  const { rowParams, item, ...restProps } = props
+
+  const rowClassName = getCardOrRowClassName(item, rowParams?.activeRowParams)
+
+  return (
+    <tr
+      {...restProps}
+      onClick={() => {
+        rowParams?.onRowClick?.(item)
+      }}
+      ref={ref as Ref<HTMLTableRowElement>}
+      className={rowClassName}
+    />
+  )
+}
+
+const TableRow = memo(
+  forwardRef(TableRowInner) as <T>(
+    props: TableRowProps<T> & { ref?: ForwardedRef<HTMLTableRowElement> },
+  ) => ReturnType<typeof TableRowInner>,
+)
+TableRow.displayName = 'TableRow'
+
+const createTableComponents = <T,>(rowParams?: TableRowParams<T>) => {
+  return {
+    //? I'm sorry:(
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    TableRow: (props: any) => <TableRow {...props} rowParams={rowParams} />,
+  }
+}
