+++ src/components/TableAntd/Table.tsx
@@ -1,28 +1,34 @@
-import { memo } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 
+import { ColumnsType } from 'antd/es/table'
 import classNames from 'classnames'
 import { isEmpty } from 'lodash'
 
 import { ViewState, useTableView } from '@banx/store'
 
 import { Loader } from '../Loader'
-import { ColumnType, SortViewParams, TableRowParams } from './types'
+import { SCROLL_THRESHOLD_HEIGHT } from './constants'
+import { ActiveRowParams, PartialBreakpoints, SortViewParams } from './types'
 import { CardView, SortView, TableView } from './views'
 
 import styles from './Table.module.less'
 
 export interface TableProps<T, P> {
   data: Array<T>
-  columns: ColumnType<T>[]
+  columns: ColumnsType<T>
+  rowKeyField: keyof T
   loading?: boolean
-  loadMore?: () => void
 
   sortViewParams?: SortViewParams<P>
-  rowParams?: TableRowParams<T> //? Must be wrapped in useMemo because of render virtual table specific
+  activeRowParams?: ActiveRowParams<T>[]
 
+  fetchMoreTrigger?: Dispatch<SetStateAction<Element | null>>
   showCard?: boolean
+  onRowClick?: (dataItem: T) => void
+  breakpoints?: PartialBreakpoints
   className?: string
   classNameTableWrapper?: string
+  scrollX?: number
   emptyMessage?: string
 }
 
@@ -30,13 +36,13 @@ const Table = <T extends object, P extends object>({
   data,
   columns,
   sortViewParams,
-  rowParams,
+  activeRowParams,
   showCard,
   loading,
   emptyMessage,
-  className,
   classNameTableWrapper,
-  loadMore,
+  fetchMoreTrigger,
+  ...props
 }: TableProps<T, P>) => {
   const { viewState } = useTableView()
 
@@ -55,14 +61,16 @@ const Table = <T extends object, P extends object>({
           <ViewComponent
             data={data}
             columns={columns}
-            rowParams={rowParams}
-            className={className}
-            loadMore={loadMore}
+            activeRowParams={activeRowParams}
+            {...props}
           />
+        )}
+        {fetchMoreTrigger && (
+          <div style={{ height: SCROLL_THRESHOLD_HEIGHT }} ref={fetchMoreTrigger} />
         )}
       </div>
     </>
   )
 }
 
-export default memo(Table) as typeof Table
+export default Table
