+++ src/components/TableAntd/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Table'
 export * from './types'
-export * from './helpers'
+export * from './Table'
