+++ src/components/TableAntd/views/CardView/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CardView'
