+++ src/components/TableAntd/views/SortView/SortView.tsx
@@ -1,26 +1,28 @@
 import { useState } from 'react'
 
+import { ColumnsType } from 'antd/es/table'
+
 import { SearchSelect, SearchSelectProps } from '@banx/components/SearchSelect'
 import { SortDropdown } from '@banx/components/SortDropdown'
 import { Toggle, ToggleProps } from '@banx/components/Toggle'
 
 import { ViewState, useTableView } from '@banx/store'
 
-import { ColumnType, SortParams } from '../../types'
+import { SortParams } from '../../types'
 import { SwitchModeButtons } from './components'
 import { parseTableColumn } from './helpers'
 
 import styles from './SortView.module.less'
 
 interface SortViewProps<T, P> {
-  columns: ColumnType<T>[]
+  columns: ColumnsType<T>
   searchSelectParams: SearchSelectProps<P>
   sortParams?: SortParams
   toggleParams?: ToggleProps
   showCard?: boolean
 }
 
-export const SortView = <T extends object, P extends object>({
+const SortView = <T extends object, P extends object>({
   columns,
   searchSelectParams,
   sortParams,
@@ -54,3 +56,5 @@ export const SortView = <T extends object, P extends object>({
     </div>
   )
 }
+
+export default SortView
