+++ src/components/TableAntd/views/SortView/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SortView'
