+++ src/components/TableAntd/views/TableView/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TableView'
