+++ src/components/TableAntd/views/index.ts
@@ -1,0 +1,3 @@
+export { default as TableView } from './TableView'
+export { default as SortView } from './SortView'
+export { default as CardView } from './CardView'
