--- src/components/TableVirtual/types.ts
@@ -1,42 +1,0 @@
-import { Key, ReactNode } from 'react'
-
-import { SearchSelectProps } from '../SearchSelect'
-import { SortDropdownProps } from '../SortDropdown'
-import { ToggleProps } from '../Toggle'
-
-export type SortParams = Omit<SortDropdownProps, 'options'>
-
-export interface SortViewParams<T> {
-  searchSelectParams: SearchSelectProps<T>
-  sortParams?: SortParams
-  toggleParams?: ToggleProps
-}
-
-export interface ColumnType<T> {
-  key: string | Key
-  title?: string | ReactNode
-  render: (record: T, key?: Key) => ReactNode
-  showSorterTooltip?: boolean //? Legacy prop
-  sorter?: boolean //? Legacy prop
-}
-
-export type TableRowActiveParams<T> = Array<{
-  field?: string
-  condition: (record: T) => boolean
-  cardClassName?: string
-  className?: string
-}>
-
-export interface TableRowParams<T> {
-  activeRowParams?: TableRowActiveParams<T>
-  onRowClick?: (dataItem: T) => void
-}
-
-export interface TableViewProps<T> {
-  data: Array<T>
-  columns: ColumnType<T>[]
-  loading?: boolean
-  loadMore?: () => void
-  rowParams?: TableRowParams<T>
-  className?: string
-}
