--- src/components/TableVirtual/views/CardView/index.ts
@@ -1,1 +1,0 @@
-export { CardView } from './CardView'
