--- src/components/TableVirtual/views/SortView/index.ts
@@ -1,1 +1,0 @@
-export { SortView } from './SortView'
