--- src/components/TableVirtual/views/TableView/index.ts
@@ -1,1 +1,0 @@
-export { TableView } from './TableView'
