--- src/components/TableVirtual/views/index.ts
@@ -1,3 +1,0 @@
-export { TableView } from './TableView'
-export { SortView } from './SortView'
-export { CardView } from './CardView'
