+++ src/pages/LendPage/components/ActivityTable/ActivityTable.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react'
 
 import EmptyList from '@banx/components/EmptyList'
-import Table from '@banx/components/TableVirtual'
+import Table from '@banx/components/Table'
 
 import { FilterTableSection } from './FilterTableSection'
 import { getTableColumns } from './columns'
