+++ src/pages/OffersPage/components/ActiveOffersTable/ActiveOffersTable.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 
 import EmptyList from '@banx/components/EmptyList'
-import Table from '@banx/components/TableVirtual'
+import Table from '@banx/components/Table'
 
 import { Loan } from '@banx/api/core'
 import { ViewState, useTableView } from '@banx/store'
