+++ src/pages/OffersPage/components/HistoryOffersTable/HistoryOffersTable.tsx
@@ -1,5 +1,5 @@
 import EmptyList from '@banx/components/EmptyList'
-import Table from '@banx/components/TableVirtual'
+import Table from '@banx/components/Table'
 
 import { Summary } from './Summary'
 import { getTableColumns } from './columns'
