+++ src/pages/OffersPage/components/PendingOffersTable/PendingOffersTable.tsx
@@ -1,5 +1,5 @@
 import EmptyList from '@banx/components/EmptyList'
-import Table from '@banx/components/TableVirtual'
+import Table from '@banx/components/Table'
 
 import { ViewState, useTableView } from '@banx/store'
