+++ src/pages/RefinancePage/components/RefinanceTable/columns.tsx
@@ -1,5 +1,5 @@
 import { HeaderCell, NftInfoCell, createSolValueJSX } from '@banx/components/TableComponents'
-import { ColumnType } from '@banx/components/TableVirtual'
+import { ColumnType } from '@banx/components/Table'
 import Timer from '@banx/components/Timer/Timer'
 
 import { Loan } from '@banx/api/core'
